Upgrade to QEMU 4.0.1. Add more shim functions to C and Python API. Added experimental Qt GUI.

+++ plugin/qemu-memory.h
@@ -25,8 +25,8 @@
 extern "C"{
 #endif
 
-bool qemu_get_virtual_memory(int cpu_id, uint64_t address, uint8_t size, uint8_t **data);
-bool qemu_set_virtual_memory(int cpu_id, uint64_t address, uint8_t size, uint8_t *data);
+bool qemu_get_virtual_memory(int cpu_id, uint64_t address, uint64_t size, uint8_t **data);
+bool qemu_set_virtual_memory(int cpu_id, uint64_t address, uint64_t size, uint8_t *data);
 void qemu_get_physical_memory(uint64_t address, uint64_t size, uint8_t **data);
 void qemu_set_physical_memory(uint64_t address, uint64_t size, uint8_t *data);
